requirements.txt
customtkinter==5.2.2
yt-dlp
pillow
requests